# old/api/docs/swagger.yaml
definitions:
...
  github_com_nicolailuther_butter_internal_dto.AdminMarketplaceCategoryDto:
    properties:
      _orders:
        type: integer
      _sellers:
        type: integer
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      service_count:
        type: integer
      total_revenue:
        type: number
      updated_at:
        type: string
    type: object
  github_com_nicolailuther_butter_internal_dto.AdminMarketplaceSellerDto:
    properties:
      bio:
        type: string
      created_at:
        type: string
      id:
        type: integer
      nickname:
        type: string
      service_count:
        type: integer
      total_orders:
        type: integer
      total_revenue:
        type: number
      updated_at:
        type: string
      user_email:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
    type: object
  github_com_nicolailuther_butter_internal_dto.AdminMarketplaceServiceDto:
    properties:
      _seller:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceSeller'
      available_spots:
        type: integer
      category_name:
        type: string
      created_at:
        type: string
      description:
        type: string
      disclaimer:
        type: string
      expectations:
        type: string
      file_id:
        type: integer
      id:
        type: integer
      image:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.File'
      marketplace_category_id:
        type: integer
      max_price:
        type: number
      min_price:
        type: number
      orders_count:
        type: integer
      platform_commission:
        type: number
      seller_id:
        type: integer
      seller_name:
        type: string
      steps_to_start:
        type: string
      terms_in:
        type: string
      title:
        type: string
      total_revenue:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  github_com_nicolailuther_butter_internal_dto.CategoryDistribution:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      orders_count:
        type: integer
      percentage:
        type: number
      revenue:
        type: number
      service_count:
        type: integer
    type: object
...
  github_com_nicolailuther_butter_internal_dto.MarketplaceAnalytics:
    properties:
      active_disputes:
        type: integer
      average_order_value:
        type: number
      completed_orders:
        type: integer
      monthly_revenue:
        type: number
      pending_orders:
        type: integer
      total_categories:
        type: integer
      total_orders:
        type: integer
      total_revenue:
        type: number
      total_sellers:
        type: integer
      total_services:
        type: integer
    type: object
...
  github_com_nicolailuther_butter_internal_dto.OrdersByStatus:
    properties:
      cancelled:
        type: integer
      completed:
        type: integer
      delivered:
        type: integer
      disputed:
        type: integer
      in_progress:
        type: integer
      paid:
        type: integer
      pending_payment:
        type: integer
      refunded:
        type: integer
    type: object
...
  github_com_nicolailuther_butter_internal_dto.RevenueByPeriod:
    properties:
      orders_count:
        type: integer
      period:
        type: string
      revenue:
        type: number
    type: object
...
  github_com_nicolailuther_butter_internal_dto.TopSeller:
    properties:
      email:
        type: string
      nickname:
        type: string
      orders_count:
        type: integer
      seller_id:
        type: integer
      total_revenue:
        type: number
    type: object
  github_com_nicolailuther_butter_internal_dto.TopService:
    properties:
      category_name:
        type: string
      orders_count:
        type: integer
      seller_name:
        type: string
      service_id:
        type: integer
      title:
        type: string
      total_revenue:
        type: number
    type: object
...
  github_com_nicolailuther_butter_internal_enums.AccountRole:
    enum:
    - main
    - backup
    type: string
    x-enum-varnames:
    - AccountRoleMain
    - AccountRoleBackup
...
  github_com_nicolailuther_butter_internal_enums.ClientIndustry:
    enum:
    - only fans
    - credit stacking
    - real estate
    - ""
    type: string
    x-enum-varnames:
    - ClientIndustryOnlyFans
    - ClientIndustryCreditStacking
    - ClientIndustryRealEstate
    - ClientUnknownIndustry
...
  github_com_nicolailuther_butter_internal_enums.Goal:
    enum:
    - dfy
    - crm
    type: string
    x-enum-comments:
      GoalCrm: Customer Relationship Management
      GoalDfy: Done For You Marketing
    x-enum-varnames:
    - GoalDfy
    - GoalCrm
  github_com_nicolailuther_butter_internal_enums.Industry:
    enum:
    - creator
    - ecommerce
    - clipper
    - coaching
    - adult_content
    type: string
    x-enum-varnames:
    - IndustryCreator
    - IndustryECommerce
    - IndustryClipper
    - IndustryCoaching
    - IndustryAdultContent
  github_com_nicolailuther_butter_internal_enums.MarketplaceOrderStatus:
    enum:
    - pending_payment
    - paid
    - payment_expired
    - waiting_info
    - in_progress
    - delivered
    - in_revision
    - completed
    - auto_completed
    - disputed
    - refunded
    - cancelled
    type: string
    x-enum-varnames:
    - MarketplaceOrderStatusPendingPayment
    - MarketplaceOrderStatusPaid
    - MarketplaceOrderStatusPaymentExpired
    - MarketplaceOrderStatusWaitingInfo
    - MarketplaceOrderStatusInProgress
    - MarketplaceOrderStatusDelivered
    - MarketplaceOrderStatusInRevision
    - MarketplaceOrderStatusCompleted
    - MarketplaceOrderStatusAutoCompleted
    - MarketplaceOrderStatusDisputed
    - MarketplaceOrderStatusRefunded
    - MarketplaceOrderStatusCancelled
...
  github_com_nicolailuther_butter_internal_enums.TeamSize:
    enum:
    - just_me
    - two_to_five
    - six_to_twenty
    - twenty_plus
    type: string
    x-enum-varnames:
    - TeamSizeJustMe
    - TeamSizeTwoToFive
    - TeamSizeSixToTwenty
    - TeamSizeTwentyPlus
  github_com_nicolailuther_butter_internal_enums.TierLevel:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      TierFree: Free tier - no subscription
      TierGrowth: Growth plan
      TierScale: Scale plan
      TierStarter: Starter plan
    x-enum-varnames:
    - TierFree
    - TierStarter
    - TierGrowth
    - TierScale
  github_com_nicolailuther_butter_internal_enums.UserRole:
    enum:
    - user
    - poster
    - admin/agency
    - client
    - team_leader
    - admin
    type: string
    x-enum-varnames:
    - UserRoleUser
    - UserRolePoster
    - UserRoleAgency
    - UserRoleClient
    - UserRoleLeader
    - UserRoleAdmin
  github_com_nicolailuther_butter_internal_enums.UserType:
    enum:
    - individual
    - agency
    type: string
    x-enum-varnames:
    - UserTypeIndividual
    - UserTypeAgency
  github_com_nicolailuther_butter_internal_models.Account:
    properties:
      account_role:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.AccountRole'
      account_url:
        type: string
      auto_generate_enabled:
        type: boolean
      auto_generate_hour:
        type: integer
      average_views:
        type: integer
      bio:
        type: string
      bio_link:
        type: string
      bio_request:
        type: string
      client_id:
        description: Should not be null
        type: integer
      created_at:
        type: string
      cross_promo:
        type: string
      daily_marketing_cost:
        type: number
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      enabled:
        type: boolean
      followers:
        type: integer
      id:
        type: integer
      last_tracked_at:
        type: string
      name:
        type: string
      onlyfans_tracking_link:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.OnlyfansTrackingLink'
      onlyfans_tracking_link_id:
        type: integer
      platform:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.Platform'
      posting_goal:
        type: integer
      profile_image:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.File'
      profile_image_id:
        type: integer
      slideshow_posting_goal:
        type: integer
      status:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.AccountStatus'
      story_posting_goal:
        type: integer
      tracking_link_url:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  github_com_nicolailuther_butter_internal_models.Client:
    properties:
      company_percentage:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      industry:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.ClientIndustry'
      name:
        type: string
      profile_image:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.File'
      profile_image_id:
        type: integer
      updated_at:
        type: string
      user_id:
        description: Should be not null
        type: integer
    type: object
...
  github_com_nicolailuther_butter_internal_models.File:
    properties:
      content_type:
        type: string
      created_at:
        type: string
      etag:
        type: string
      id:
        type: integer
      path:
        type: string
      size:
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  github_com_nicolailuther_butter_internal_models.MarketplaceCategory:
    properties:
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_nicolailuther_butter_internal_models.MarketplaceOrder:
    properties:
      _buyer:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.User'
      _category:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceCategory'
      _client:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Client'
      _package:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceServicePackage'
      _service:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceService'
      auto_completion_date:
        type: string
      buyer_id:
        type: integer
      category_id:
        type: integer
      client_id:
        type: integer
      completed_at:
        type: string
      conversation_id:
        type: integer
      created_at:
        type: string
      days_extended:
        type: integer
      due_date:
        type: string
      id:
        type: integer
      payment:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Payment'
      payment_id:
        type: integer
      required_info:
        type: string
      revisions_used:
        type: integer
      service_id:
        type: integer
      service_package_id:
        type: integer
      start_date:
        type: string
      status:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.MarketplaceOrderStatus'
      total_deliveries:
        type: integer
      total_revisions:
        type: integer
      updated_at:
        type: string
    type: object
...
  github_com_nicolailuther_butter_internal_models.MarketplaceSeller:
    properties:
      bio:
        type: string
      created_at:
        type: string
      id:
        type: integer
      nickname:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  github_com_nicolailuther_butter_internal_models.MarketplaceService:
    properties:
      _seller:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceSeller'
      available_spots:
        type: integer
      created_at:
        type: string
      description:
        type: string
      disclaimer:
        type: string
      expectations:
        type: string
      file_id:
        type: integer
      id:
        type: integer
      image:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.File'
      marketplace_category_id:
        type: integer
      platform_commission:
        type: number
      seller_id:
        type: integer
      steps_to_start:
        type: string
      terms_in:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  github_com_nicolailuther_butter_internal_models.MarketplaceServicePackage:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration_days:
        type: integer
      id:
        type: integer
      included_revisions:
        type: integer
      name:
        type: string
      price:
        type: number
      service_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  github_com_nicolailuther_butter_internal_models.Payment:
    properties:
      amount:
        description: Amount in smallest currency unit (cents)
        type: integer
      completed_at:
        type: string
      created_at:
        type: string
      currency:
        type: string
      failure_code:
        type: string
      failure_reason:
        description: Error handling
        type: string
      id:
        type: integer
      internal_notes:
        description: Notes for admin/internal use
        type: string
      metadata:
        description: Additional metadata (stored as JSON for flexibility)
        type: string
      payment_method_last4:
        description: Last 4 digits of card, or crypto address suffix
        type: string
      payment_method_type:
        allOf:
        - $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.PaymentMethodType'
        description: Payment method information (card, crypto, etc. - all processed
          via Stripe)
      payment_type:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.PaymentType'
      payment_url:
        description: Generic payment URL (for checkout pages)
        type: string
      processed_at:
        description: Transaction details
        type: string
      provider:
        allOf:
        - $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.PaymentProvider'
        description: Core payment information
      refund_amount:
        type: integer
      refunded_at:
        type: string
      related_entity_id:
        type: integer
      related_entity_type:
        description: e.g., "marketplace_order", "subscription"
        type: string
      status:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.PaymentStatus'
      stripe_charge_id:
        type: string
      stripe_checkout_session_id:
        description: Stripe fields (all payments use Stripe)
        type: string
      stripe_customer_id:
        type: string
      stripe_invoice_id:
        type: string
      stripe_payment_intent_id:
        type: string
      updated_at:
        type: string
      user_id:
        description: User/Entity relationships
        type: integer
    type: object
  github_com_nicolailuther_butter_internal_models.PaymentMethodType:
    enum:
    - card
    - crypto
    - unknown
    type: string
    x-enum-varnames:
    - PaymentMethodCard
    - PaymentMethodCrypto
    - PaymentMethodUnknown
  github_com_nicolailuther_butter_internal_models.PaymentProvider:
    enum:
    - stripe
    type: string
    x-enum-varnames:
    - PaymentProviderStripe
...
  github_com_nicolailuther_butter_internal_models.PaymentType:
    enum:
    - marketplace_order
    - subscription
    type: string
    x-enum-varnames:
    - PaymentTypeMarketplaceOrder
    - PaymentTypeSubscription
...
  github_com_nicolailuther_butter_internal_models.User:
    properties:
      assigned_accounts:
        items:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Account'
        type: array
      assigned_clients:
        items:
          $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Client'
        type: array
      changed_password:
        type: boolean
      created_at:
        type: string
      created_by:
        type: integer
      current_period_end:
        type: string
      email:
        type: string
      goal:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.Goal'
      id:
        type: integer
      industry:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.Industry'
      name:
        type: string
      password:
        type: string
      referral_code_id:
        type: integer
      role:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.UserRole'
      stripe_customer_id:
        description: Stripe-related fields for subscription management
        type: string
      subscription_status:
        type: string
      team_size:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.TeamSize'
      tier_level:
        allOf:
        - $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.TierLevel'
        description: User's subscription tier level
      type:
        allOf:
        - $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.UserType'
        description: Onboarding preferences
      updated_at:
        type: string
    type: object
  github_com_nicolailuther_butter_internal_responses.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact:
    name: Default
    url: https://default.dev
  description: Backend service for Butter
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Butter API
  version: "1.0"
paths:
...
  /api/v1/admin/marketplace/analytics:
    get:
      description: Retrieve marketplace analytics including total revenue, orders,
        sellers, and services stats
      produces:
      - application/json
      responses:
        "200":
          description: Marketplace analytics data
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.MarketplaceAnalytics'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get marketplace analytics dashboard data
      tags:
      - admin-marketplace
  /api/v1/admin/marketplace/analytics/category-distribution:
    get:
      description: Retrieve distribution of orders and revenue by category
      produces:
      - application/json
      responses:
        "200":
          description: Category distribution data
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.CategoryDistribution'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get marketplace category distribution
      tags:
      - admin-marketplace
  /api/v1/admin/marketplace/analytics/orders-by-status:
    get:
      description: Retrieve count of marketplace orders grouped by status
      produces:
      - application/json
      responses:
        "200":
          description: Orders count by status
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.OrdersByStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get marketplace orders by status
      tags:
      - admin-marketplace
  /api/v1/admin/marketplace/analytics/revenue:
    get:
      description: Retrieve marketplace revenue grouped by day, week, or month
      parameters:
      - default: month
        description: Period (day, week, month)
        in: query
        name: period
        type: string
      - default: 30
        description: Number of days to look back
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Revenue data by period
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.RevenueByPeriod'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get marketplace revenue by period
      tags:
      - admin-marketplace
  /api/v1/admin/marketplace/analytics/top-sellers:
    get:
      description: Retrieve top performing sellers by revenue or order count
      parameters:
      - default: 10
        description: Number of sellers to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Top sellers list
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.TopSeller'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get top marketplace sellers
      tags:
      - admin-marketplace
  /api/v1/admin/marketplace/analytics/top-services:
    get:
      description: Retrieve top performing services by revenue or order count
      parameters:
      - default: 10
        description: Number of services to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Top services list
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.TopService'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get top marketplace services
      tags:
      - admin-marketplace
  /api/v1/admin/marketplace/categories:
    get:
      description: Retrieve all marketplace categories with service counts and order
        stats
      produces:
      - application/json
      responses:
        "200":
          description: List of categories with analytics
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.AdminMarketplaceCategoryDto'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all marketplace categories with admin details
      tags:
      - admin-marketplace
  /api/v1/admin/marketplace/orders:
    get:
      description: Retrieve all marketplace orders with buyer, seller, and service
        details
      produces:
      - application/json
      responses:
        "200":
          description: List of all orders
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.MarketplaceOrder'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all marketplace orders with admin details
      tags:
      - admin-marketplace
  /api/v1/admin/marketplace/sellers:
    get:
      description: Retrieve all marketplace sellers with service counts and order
        stats
      produces:
      - application/json
      responses:
        "200":
          description: List of sellers with analytics
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.AdminMarketplaceSellerDto'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all marketplace sellers with admin details
      tags:
      - admin-marketplace
  /api/v1/admin/marketplace/services:
    get:
      description: Retrieve all marketplace services with order stats
      produces:
      - application/json
      responses:
        "200":
          description: List of services with analytics
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.AdminMarketplaceServiceDto'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all marketplace services with admin details
      tags:
      - admin-marketplace
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  PushApiKey:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"